+++ History.md
@@ -1,5 +1,6 @@
 2.0.2 / 2017-08-??
 ==================
+  * Fix instrumentation for ES6 class members (https://github.com/tntim96/JSCover/issues/261)
   * Fix instrumentation for array definition across lines (https://github.com/tntim96/JSCover/issues/260)
 
 2.0.1 / 2017-07-23